Make all expense route protected

+++ src/controller/ExpenseController.ts
@@ -2,10 +2,11 @@ import { Request, Response } from "express";
 import Joi from "joi";
 import { CustomReq } from "../config/CustomReq";
 import { ExpenseDto } from "../dto/ExpenseDto";
+import { AuthMiddlewareReq } from "../dto/UserDto";
 import { Expense } from "../models/ExpenseModel";
 import { APIFilter } from "../utils/API_Filter";
 
-export const getExpense = async (req: Request, res: Response) => {
+export const getExpense = async (req: AuthMiddlewareReq, res: Response) => {
   const typedReqQuery = req.query as { [key: string]: string };
   const apiFilterInstance = new APIFilter(Expense.find(), typedReqQuery, [
     "title",
@@ -20,7 +21,7 @@ export const getExpense = async (req: Request, res: Response) => {
     query = query.find({ title: { $in: title } });
   }
 
-  const expense = await query;
+  const expense = await query.find({ user: req.user?._id });
   res.status(200).json(expense);
 };
 
@@ -37,6 +38,7 @@ export const saveExpense = async (
     description: req.body.description,
     category: req.body.category,
     date: req.body.date,
+    user: req.body.user,
   });
 
   res.status(201).json(savedExpense);
@@ -48,6 +50,12 @@ export const updateExpense = async (
 ) => {
   const { error } = validateExpense(req.body);
   if (error) return res.status(400).json({ message: error.details[0].message });
+  if (!req.user) return res.status(404).json({ message: "User not found" });
+  const expense = await Expense.findById(req.params.id);
+  if (!expense) return res.status(404).json({ message: "Expense not found" });
+
+  if (expense.user.toString() !== req.user._id?.toString())
+    return res.status(404).json({ message: "User not found" });
   const updatedExpense = await Expense.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -63,6 +71,12 @@ export const deleteExpense = async (
   req: CustomReq<ExpenseDto>,
   res: Response
 ) => {
+  if (!req.user) return res.status(404).json({ message: "User not found" });
+  const expense = await Expense.findById(req.params.id);
+  if (!expense) return res.status(404).json({ message: "Expense not found" });
+
+  if (req.user._id?.toString() !== expense.user.toString())
+    return res.status(404).json({ message: "User not found" });
   const deletedExpense = await Expense.findByIdAndDelete<ExpenseDto>(
     req.params.id
   );
@@ -78,6 +92,7 @@ const validateExpense = (expense: ExpenseDto) => {
     amount: Joi.number().required(),
     description: Joi.string(),
     date: Joi.date(),
+    user: Joi.string().required(),
   });
 
   return schema.validate(expense);
